Initial Commit of the Data Display App, No Server Code yet.
This is the base app with some functions added to fake data updates. We
need to write the node.js server, and hook the server into the app. And
fix the freaking window.resize thing to be an on click action for nav
bar items. Waiting for the interval functions to cause the redraw annoys
me a LOT.