Updated gyroscope chart to be prettier, started modular design for updating charts
I started several functions to isolate the fake update code so that it
will be easy to replace when we have real code/data to take its place.
The gyro is now an angular gauge instead of a polar chart due to issues
with polar chart capabilities (no straight (vertical) lines). Added
exporting.js to allow for saving images if the user wants.